+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotatedVCFWriter.java
@@ -42,7 +42,7 @@ public class AnnotatedVCFWriter extends AnnotatedVariantWriter {
 		this.refDict = refDict;
 		this.annotator = new VariantContextAnnotator(refDict, chromosomeMap, new VariantContextAnnotator.Options(
 				InfoFields.build(options.writeVCFAnnotationStandardInfoFields, options.writeJannovarInfoFields),
-				!options.showAll));
+				!options.showAll, options.escapeAnnField));
 		this.vcfPath = vcfPath;
 		this.options = options;