+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/JannovarOptions.java
@@ -67,6 +67,9 @@ public final class JannovarOptions {
 	/** whether or not to write out the VCF annotation standard fields */
 	public boolean writeVCFAnnotationStandardInfoFields = true;
 
+	/** whether or not to escape characters like <code>'='</code> in the <code>ANN</code> VCF field. */
+	public boolean escapeAnnField = true;
+
 	// TODO(holtgrem): enable and use this!
 	/** path to output VCF file path (overrides generation of file name from input file name) */
 	public String outVCFFile = null;