+++ jannovar-htsjdk/src/main/java/de/charite/compbio/jannovar/htsjdk/VariantContextAnnotator.java
@@ -38,6 +38,9 @@ public final class VariantContextAnnotator {
 		/** selection of info fields to write out (defaults to {@link InfoFields#VCF_ANN}) */
 		public final InfoFields infoFields;
 
+		/** whether or not to escape values in the ANN field (defaults to <code>true</code>) */
+		public final boolean escapeAnnField;
+
 		/**
 		 * Whether or not to trim each annotation list to the first (one with highest putative impact), defaults to
 		 * <code>true</code>
@@ -47,11 +50,13 @@ public final class VariantContextAnnotator {
 		public Options() {
 			infoFields = InfoFields.VCF_ANN;
 			oneAnnotationOnly = true;
+			escapeAnnField = true;
 		}
 
-		public Options(InfoFields infoFields, boolean oneAnnotationOnly) {
+		public Options(InfoFields infoFields, boolean oneAnnotationOnly, boolean escapeAnnField) {
 			this.infoFields = infoFields;
 			this.oneAnnotationOnly = oneAnnotationOnly;
+			this.escapeAnnField = escapeAnnField;
 		}
 	}